requirements.txt
python-dotenv
pymongo[srv]>=3.11
langchain
langchain-community
langchain-text-splitters
pymupdf
pinecone
ipykernel
certifi
langchain-groq
numpy
langsmith
sentence-transformers